pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flashback-screenshots"
version = "0.1.0"
description = "Search through your screenshot history with OCR and semantic search"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "James Brown"}
]
keywords = ["screenshot", "ocr", "search", "history", "flashback", "semantic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
    "Environment :: X11 Applications",
]

dependencies = [
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "platformdirs>=3.0.0",
    "psutil>=5.9.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]
embedding = ["requests>=2.28.0"]
search = ["nltk>=3.8.0", "rank-bm25>=0.2.2", "jieba>=0.42.1"]
chinese = ["jieba>=0.42.1"]
webui = ["fastapi>=0.100.0", "uvicorn>=0.23.0", "jinja2>=3.1.0", "python-multipart>=0.0.6", "aiofiles>=23.0.0"]
linux = ["python-xlib>=0.33"]
windows = ["pygetwindow>=0.0.9", "pywin32>=306"]
all = ["flashback-screenshots[ocr,search,webui]"]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
flashback = "flashback.cli.main:main"
flashback-api = "flashback.api.server:main"

[project.urls]
Homepage = "https://github.com/james4ever0/flashback-screenshots"
Repository = "https://github.com/james4ever0/flashback-screenshots"
Issues = "https://github.com/james4ever0/flashback-screenshots/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["flashback*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
